+++ CHANGELOG.md
@@ -70,6 +70,7 @@ All notable changes to this project will be documented in this file.
 - Add Component package manager support - component.json
 
 ### Fixed
+- updating a value on onincomplete event doesn't work #955
 - $.inputmask.isValid("1A", { mask : "1A" }) returns false #858
 - IE8 doesn't support window.getSelection js error #853
 - Email with dot - paste not working #847